requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
greenlet==2.0.2
numpy==1.24.3
pandas==2.0.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
whitenoise==6.2.0